Rework job creation logic to not create new jobs on printer restart.
This change removes a lot of the code which would put the job into a
wait list until the client inits, however this is unnecessary, because
the currentJob property persists, so we kick off right where we started.
Some of the event namings and layouts were also reworked due to the
usage of them requiring a lot of boilerplate.